fix(metro-resolver-symlinks): type info is empty
