+++ change/@rnx-kit-metro-resolver-symlinks-20b3abcb-3c1a-4437-9774-1560f69b2ac5.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "Fix type information not being generated correctly",
+  "packageName": "@rnx-kit/metro-resolver-symlinks",
+  "email": "4123478+tido64@users.noreply.github.com",
+  "dependentChangeType": "patch"
+}
