+++ packages/test-app/metro.config.js
@@ -1,4 +1,5 @@
 const { exclusionList, makeMetroConfig } = require("@rnx-kit/metro-config");
+const MetroSymlinksResolver = require("@rnx-kit/metro-resolver-symlinks");
 
 // Metro will pick up react-native-macos/-windows mocks if we don't exclude them
 const blockList = exclusionList([/.*__fixtures__.*/]);
@@ -8,5 +9,6 @@ module.exports = makeMetroConfig({
   resolver: {
     blacklistRE: blockList,
     blockList,
+    resolveRequest: MetroSymlinksResolver(),
   },
 });
